pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rosetta"
version = "1.0.0"
description = "Cross-DBMS SQL behavioral consistency verification tool"
requires-python = ">=3.8"
dependencies = [
    "pymysql>=1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
mysql-connector = ["mysql-connector-python>=8.0"]

[project.scripts]
rosetta = "rosetta.cli:main"

[tool.setuptools.packages.find]
where = ["."]
